+++ src/window_type.h
@@ -813,7 +813,7 @@ struct Window;
 struct WindowBase;
 
 /** Number to differentiate different windows of the same class */
-typedef int32 WindowNumber;
+typedef int64 WindowNumber;
 
 /** State of handling an event. */
 enum EventState {
